Add files via upload
Working on final version of study code. This branch has most of the structure built and established. Still needs the Melodic signals developed.

Note: I'm not including the DLL, as I'm unsure of including the actual material provided for interfacing with my equipment. So this code isn't directly functional on its own, but the overall design and framework is instructive for how to manipulate a haptics system